+++ app/router/middleware.go
@@ -1,0 +1,4 @@
+package router
+
+
+//TODO create middleware
